pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "edge-proxy-py"
version = "0.1.1"
description = "Edge/IoT adaptive reverse proxy with policy routing, caching, and offline-aware fallback."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Kubenew"}]
keywords = ["edge", "iot", "reverse-proxy", "caching", "routing", "asyncio"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: System :: Networking",
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "typer>=0.12.3",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.27.0",
]
dev = ["edge-proxy-py[test]", "ruff>=0.1"]

[project.scripts]
edge-proxy = "edge_proxy_py.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["edge_proxy_py"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
